Upgrade to version 1.6.0
The main issues included in this release are:

 PubSub Server / API REST:

 1. Possibility to reject events associated with both no registered and disabled sensors

   Added new configuration parameter that allows configure if a request associated with a no registered sensor (or disabled) should be rejected. In previous versions these requests were not rejected.

 2. Possibility to reject events associated with both no registered and disabled alerts

   Idem to previous issue but applied to alerts.

 3. Notification retry

   New retry feature that allows future retry if remote endpoint is unreachable. As each client have different requirements, retry parameters (number of retries and delay between each one) are defined in each subscription.

 4. New HTTPs access control

   Now it is possible to configure (on the Catalog) that requests from a provider (or an application client) should be secured with SSL. If an entity has the HTTPs flag configured then non-secure requests will be rejected.

 5. New fields in notifications sent by Sentilo.

  Notifications now include four new fields: publisher (the entity who has made the request), publishedAt (the time at which the event arrived at Sentilo), publisherTenant and tenant (both applies only in a  multi-tenant instance)

 Catalog:

 6. Google API key parameter

   As of June 22, 2016 Google Maps V3 no longer supports keyless access so you need to get a key for every (referrer-)domain which has never had a Google Map on it.
   Due to this reason, we have added a new configuration parameter that allows to set this key on your Sentilo instance.

 7. Sensor state / substate

   New field state that allows enable / disable sensors. A disabled sensor has the same behaviour as an unregistered sensor.
   New field substate allows to add a description to the state.

 8. Alert state

   New field state that allows enable / disable alerts.

 9. Bulk alert generator

   New feature that allows to create massive alerts on Catalog based on user-defined criteria

 Agents:

 10. Relational agent retry backend

   New feature that allows future retry if remote database is unreachable.

 11. New Elasticsearch agent

   New Sentilo agent that lets you index events directly into Elasticsearch from Sentilo.

 12. New openTSDB agent

   New Sentilo agent that lets you store your events directly into openTSDB.

+++ sentilo-catalog-web/src/main/java/org/sentilo/web/catalog/service/impl/ProviderServiceImpl.java
@@ -42,7 +42,6 @@ import org.sentilo.web.catalog.domain.Alert;
 import org.sentilo.web.catalog.domain.Component;
 import org.sentilo.web.catalog.domain.Provider;
 import org.sentilo.web.catalog.domain.Sensor;
-import org.sentilo.web.catalog.event.DeletePlatformResourcesEvent;
 import org.sentilo.web.catalog.repository.ProviderRepository;
 import org.sentilo.web.catalog.search.SearchFilter;
 import org.sentilo.web.catalog.service.PermissionService;
@@ -55,10 +54,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
 @Service
-public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> implements ProviderService {
+public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider>implements ProviderService {
 
   @Autowired
   private ProviderRepository repository;
@@ -95,7 +95,7 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.sentilo.web.catalog.service.ProviderService#findAllowed()
    */
   @Override
@@ -110,20 +110,18 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see
-   * org.sentilo.web.catalog.service.ProviderService#deleteChildDocuments(org.sentilo.web.catalog
-   * .domain.Provider)
+   * org.sentilo.web.catalog.service.ProviderService#deleteChildrens(org.sentilo.web.catalog.domain.
+   * Provider)
    */
-  public void deleteChilds(final Provider provider) {
-    notifyProviderToDelete(provider);
-    deleteRelatedResources(provider);
-    tenantPermissionService.deleteRelatedEntity(provider);
+  public void deleteChildren(final Provider provider) {
+    deleteRelatedResources(provider, false);
   }
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.sentilo.web.catalog.service.ProviderService#addGrantedTenant(java.lang.String,
    * java.lang.String)
    */
@@ -136,7 +134,7 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.sentilo.web.catalog.service.ProviderService#removeGrantedTenant(java.lang.String,
    * java.lang.String)
    */
@@ -149,7 +147,7 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.sentilo.web.catalog.service.ProviderService#getGrantedTenantsIds(java.lang.String)
    */
   @Override
@@ -159,7 +157,7 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.sentilo.web.catalog.service.ProviderService#deleteFromTenant(java.lang.String)
    */
   @Override
@@ -168,12 +166,14 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
     filter.addAndParam("tenantId", tenantId);
     final Query query = buildQuery(filter);
     final List<Provider> providers = getMongoOps().find(query, Provider.class);
-    delete(providers);
+    if (!CollectionUtils.isEmpty(providers)) {
+      delete(providers);
+    }
   }
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see org.sentilo.web.catalog.service.ProviderService#isProviderFromTenant(java.lang.String,
    * java.lang.String)
    */
@@ -185,7 +185,7 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see
    * org.sentilo.web.catalog.service.impl.AbstractBaseCrudServiceImpl#doBeforeCreate(org.sentilo
    * .web.catalog.domain.CatalogDocument)
@@ -197,7 +197,7 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
       provider.setName(provider.getId());
     }
 
-    // Validamos la unicidad del identificador: no puede sexistir otra entidad (app o provider) con
+    // Validamos la unicidad del identificador: no puede existir otra entidad (app o provider) con
     // el mismo identificador.
     checkIntegrityKey(provider.getId());
 
@@ -208,51 +208,49 @@ public class ProviderServiceImpl extends AbstractBaseCrudServiceImpl<Provider> i
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see
    * org.sentilo.web.catalog.service.impl.AbstractBaseCrudServiceImpl#doAfterDelete(org.sentilo.
    * web.catalog.domain.CatalogDocument)
    */
   protected void doAfterDelete(final Provider provider) {
-    permissionService.deleteRelated(provider);
-    deleteChilds(provider);
+    doAfterDelete(Arrays.asList(new Provider[] {provider}));
   }
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see
    * org.sentilo.web.catalog.service.impl.AbstractBaseCrudServiceImpl#delete(java.util.Collection)
    */
   protected void doAfterDelete(final Collection<Provider> providers) {
-    notifyProvidersToDelete(providers);
     for (final Provider provider : providers) {
-      permissionService.deleteRelated(provider);
       deleteRelatedResources(provider);
     }
   }
 
   private void deleteRelatedResources(final Provider provider) {
-    // Must be remove all resources related to the provider: alerts, sensors and components
+    deleteRelatedResources(provider, true);
+  }
+
+  private void deleteRelatedResources(final Provider provider, final boolean fullDelete) {
+    // Should be removed all resources related to the provider: alerts, sensors, components,
+    // permissions and tenantPermissions (these resources only in the case of a full delete)
     final SearchFilter childFilter = new SearchFilter();
     childFilter.addAndParam("providerId", provider.getId());
     final Query childQuery = buildQuery(childFilter);
 
-    getMongoOps().remove(childQuery, Alert.class);
-    getMongoOps().remove(childQuery, Sensor.class);
-    getMongoOps().remove(childQuery, Component.class);
+    doDelete(childQuery, Alert.class);
+    doDelete(childQuery, Sensor.class);
+    doDelete(childQuery, Component.class);
 
-    // Also delete all tenant permissions related with it
-    tenantPermissionService.deleteRelatedEntity(provider);
-  }
+    if (fullDelete) {
+      // Also delete all tenant permissions related to it
+      tenantPermissionService.deleteRelatedEntity(provider);
 
-  private void notifyProviderToDelete(final Provider provider) {
-    notifyProvidersToDelete(Arrays.asList(provider));
-  }
-
-  private void notifyProvidersToDelete(final Collection<Provider> providers) {
-    // Create new DeletePlatformResourceEvent to notify what providers will be deleted
-    getContext().publishEvent(new DeletePlatformResourcesEvent<Provider>(this, providers, Provider.class));
+      // And finally remove related permissions over other entities
+      permissionService.deleteRelated(provider);
+    }
   }
 
 }
